pyproject.toml
[tool.poetry]
name = "absa-llm"
version = "0.1.0"
description = ""
authors = ["Paul Simmering <paul.simmering@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
pandas = "^1.5.3"
openai = "^0.28.0"
python-dotenv = "^1.0.0"
jsonlines = "^3.1.0"
tqdm = "^4.65.0"
tiktoken = "^0.3.2"
pyarrow = "^11.0.0"
confection = "^0.1.0"
pyprojroot = "^0.3.0"
xmltodict = "^0.13.0"
petname = "^2.6"
jsonschema = "4.18.0"
plotnine = "^0.12.2"
tabulate = "^0.9.0"
pins = "^0.8.3"
s3fs = "2023.5.0"
boto3 = "1.26.145"
pytest = "^7.4.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.21.3"
nbformat = "^5.8.0"
nbclient = "^0.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
